chore: clean up new repositories
